Convert FakeOutlineTunnel to async/await; rename file
