+++ src/www/app/cordova_main.ts
@@ -22,7 +22,7 @@ import {EventQueue} from '../model/events';
 import {AbstractClipboard, Clipboard, ClipboardListener} from './clipboard';
 import {EnvironmentVariables} from './environment';
 import {SentryErrorReporter} from './error_reporter';
-import {FakeOutlineTunnel} from './fake_connection';
+import {FakeOutlineTunnel} from './fake_tunnel';
 import {main} from './main';
 import {OutlineServer} from './outline_server';
 import {OutlinePlatform} from './platform';
